(#113) Fixed an issue where GetLatest wouldn't actually get the latests.
Progress is not reported correctly when installing packages that
download their installers.
Async all the way down!
Added Package Update All button.

+++ Chocolatey.Gui/Services/IProgressService.cs
@@ -7,7 +7,7 @@ using MahApps.Metro.Controls.Dialogs;
 
 namespace Chocolatey.Gui.Services
 {
-    public interface IProgressService : INotifyPropertyChanged, IProgress<int>
+    public interface IProgressService : INotifyPropertyChanged, IProgress<double>
     {
         bool IsLoading { get; }
         void StartLoading(string title = "", string message = "");
